+++ server/src/main/java/luan/moonvs/models/builders/ContentBuilder.java
@@ -2,6 +2,7 @@ package luan.moonvs.models.builders;
 
 import luan.moonvs.models.entities.Content;
 import luan.moonvs.models.enums.ContentType;
+import luan.moonvs.models.tmdb_responses.TmdbContent;
 import luan.moonvs.models.tmdb_responses.TmdbGenres;
 import luan.moonvs.models.tmdb_responses.TmdbMovie;
 import luan.moonvs.models.tmdb_responses.TmdbTv;
@@ -25,6 +26,27 @@ public class ContentBuilder {
         return contentBuilder;
     }
 
+    public static ContentBuilder create(TmdbContent content, ContentType contentType) {
+        ContentBuilder contentBuilder = new ContentBuilder();
+        contentBuilder.content = new Content();
+
+        contentBuilder.content.setIdTmdb(content.id());
+        contentBuilder.content.setAdult(content.adult());
+        contentBuilder.content.setOriginalTitle(content.originalTitle());
+        contentBuilder.content.setPtTitle(content.ptTitle());
+        contentBuilder.content.setOverview(content.overview());
+        contentBuilder.content.setGenres(content.genres().stream()
+                .map(TmdbGenres::genre)
+                .collect(Collectors.toList()));
+        contentBuilder.content.setPosterPath(content.posterPath());
+        contentBuilder.content.setTmdbVoteAvg(content.voteAvg());
+        contentBuilder.content.setTmdbVoteCount(content.voteCount());
+        contentBuilder.content.setContentType(contentType);
+
+        return contentBuilder;
+    }
+
+    @Deprecated
     public static ContentBuilder create(TmdbMovie tmdbMovie) {
         ContentBuilder contentBuilder = new ContentBuilder();
         contentBuilder.content = new Content();
@@ -45,6 +67,7 @@ public class ContentBuilder {
         return contentBuilder;
     }
 
+    @Deprecated
     public static ContentBuilder create(TmdbTv tmdbTv) {
         ContentBuilder contentBuilder = new ContentBuilder();
         contentBuilder.content = new Content();