+++ server/src/main/java/luan/moonvs/models/tmdb_responses/TmdbTv.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.util.List;
 
+@Deprecated
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record TmdbTv(int id,
                      List<TmdbGenres> genres,