+++ server/src/main/java/luan/moonvs/services/TmdbService.java
@@ -1,16 +1,17 @@
 package luan.moonvs.services;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.module.SimpleModule;
 import jakarta.annotation.PostConstruct;
 import luan.moonvs.models.builders.ContentBuilder;
 import luan.moonvs.models.entities.Content;
-import luan.moonvs.models.tmdb_responses.TmdbMovie;
-import luan.moonvs.models.tmdb_responses.TmdbResults;
-import luan.moonvs.models.tmdb_responses.TmdbSearch;
-import luan.moonvs.models.tmdb_responses.TmdbTv;
+import luan.moonvs.models.enums.ContentType;
+import luan.moonvs.models.tmdb_responses.*;
 import luan.moonvs.models.tmdb_responses.providers.ProviderResults;
 import luan.moonvs.models.tmdb_responses.providers.ProviderType;
 import luan.moonvs.utils.HttpRequestEntity;
 import luan.moonvs.utils.RequestEntity;
+import luan.moonvs.utils.deserializers.TmdbContentDeserializer;
 import org.apache.http.client.utils.URIBuilder;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
@@ -36,6 +37,7 @@ class TmdbService {
     private HttpHeaders HEADERS;
 
     private final int MAX_PAGES = 50;
+    private final String API_VERSION = "3";
 
     @PostConstruct
     private void init() {
@@ -57,8 +59,7 @@ class TmdbService {
     }
 
     public List<TmdbSearch> search(String title, boolean includeAdult) throws URISyntaxException {
-        final String URL = String.format(newUrl, "search/multi");
-
+        final String type = "multi";
         final String language = "en-US";
         int page = 1;
         Map<String, String> parameters = new LinkedHashMap<>() {{
@@ -68,19 +69,16 @@ class TmdbService {
         }};
         List<TmdbSearch> results = new ArrayList<>();
 
-        while(true) {
+        while (true) {
             parameters.put("page", String.valueOf(page));
-            RequestEntity<?, TmdbResults> request = new RequestEntity<>(
-                    URL,
-                    HttpMethod.GET,
-                    HEADERS,
-                    TmdbResults.class
-            );
-            request.addParameters(parameters);
-            ResponseEntity<TmdbResults> searchResults = request.exchange();
+            var request = new RequestEntity<>(newUrl, HttpMethod.GET, HEADERS, TmdbResults.class);
+            var searchResults = request
+                    .setPath(API_VERSION, "search", type)
+                    .addParameters(parameters)
+                    .exchange();
 
             if (!searchResults.getStatusCode().is2xxSuccessful()
-            ||  !searchResults.hasBody())
+                    || !searchResults.hasBody())
                 return results;
 
             TmdbResults response = searchResults.getBody();
@@ -94,6 +92,7 @@ class TmdbService {
         }
     }
 
+    @Deprecated
     public Content viewMovie(int id) {
         final String QUERY = "movie/";
         final String URL = String.format(baseUrl, QUERY, id);
@@ -117,6 +116,7 @@ class TmdbService {
                 .build();
     }
 
+    @Deprecated
     public Content viewSeries(int id) {
         final String QUERY = "tv/";
         final String URL = String.format(baseUrl, QUERY, id);
@@ -140,6 +140,31 @@ class TmdbService {
                 .build();
     }
 
+    public Content getContent(int id, ContentType contentType) throws URISyntaxException {
+        final String language = "en-US";
+
+        final var request = new RequestEntity<>(newUrl, HttpMethod.GET, HEADERS, TmdbContent.class);
+        final var content = request
+                .setPath(API_VERSION, contentType.getContentType(), String.valueOf(id))
+                .addParameter("language", language)
+                .exchange();
+
+        if (!content.getStatusCode().is2xxSuccessful()
+                || !content.hasBody())
+            return null;
+
+        ProviderResults providers = getWatchProviders(contentType, id);
+        ProviderType brProviders = null;
+        if (providers != null)
+            brProviders = providers.results().get("BR");
+
+        return ContentBuilder
+                .create(content.getBody(), contentType)
+                .withProviders(brProviders)
+                .build();
+    }
+
+    @Deprecated
     private ProviderResults getWatchProviders(final String QUERY, int id) {
         final String URL = String.format(baseUrl, QUERY, id + "/watch/providers");
 
@@ -147,6 +172,19 @@ class TmdbService {
         ResponseEntity<ProviderResults> providers = request.exchange();
 
         if (!providers.getStatusCode().is2xxSuccessful())
+            return null;
+
+        return providers.getBody();
+    }
+
+    private ProviderResults getWatchProviders(final ContentType contentType, int id) throws URISyntaxException {
+        final var request = new RequestEntity<>(newUrl, HttpMethod.GET, HEADERS, ProviderResults.class);
+        final var providers = request
+                .setPath(API_VERSION, contentType.getContentType(), String.valueOf(id), "watch", "providers")
+                .exchange();
+
+        if (!providers.getStatusCode().is2xxSuccessful()
+                || !providers.hasBody())
             return null;
 
         return providers.getBody();